requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
icalendar>=5.0.11
selenium>=4.15.0
python-dateutil>=2.8.2
PyPDF2>=3.0.0
pdfplumber>=0.10.0
